feat: editable document review UI + automatic expiry-reminder scheduler
Document Review UI (titan-ui):
- Extraction table is now editable — korisnik ispravlja opis/količinu/jedinicu
  prije generiranja ponude, plus 🗑 za uklanjanje krivo ekstrahirane stavke.
- createQuoteFromDoc skuplja ispravke u selected_items (zadržava catalog match)
  i šalje na /create-quote (backend već podržava selected_items).

In-process scheduler (titan-core / api-architect):
- app/services/scheduler.py: asyncio background task pokrenut u lifespanu —
  bez Celery/brokera (solo-dev čisti uvicorn). Podsjetnici za ponude koje
  uskoro ističu (status=sent, valid_until ≤ 3 dana), grupirani po org,
  šalje vlasniku/adminu. Dedup preko audit zapisa 'quote.reminder_sent'.
- ENABLE_SCHEDULER config flag; ne diže se u testovima ni bez SMTP-a.

Tests (qa-war-machine): +3 integration testa (doc review edit/fallback,
scheduler send+dedup). 37/37 prolazi. Lint 0 (ignore RUF001-003 za hrv. tekst).

+++ backend/app/api/v1/analytics.py
@@ -397,7 +397,7 @@ async def client_profile(
     score += min(total_won_value / 1000, 100) * 0.3  # €100k+ → max
     if avg_margin:
         score += min(avg_margin / 30 * 100, 100) * 0.3
-    score_int = int(round(score))
+    score_int = round(score)
     rating = "A" if score_int >= 70 else "B" if score_int >= 45 else "C" if score_int >= 25 else "D"
 
     notes: list[str] = []
